Split ChannelList into two seperate files
