+++ OutlookMatters/Mattermost/Channel.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace OutlookMatters.Mattermost
+{
+    public class Channel
+    {
+        [JsonProperty("id")]
+        public string ChannelId { get; set; }
+
+        [JsonProperty("display_name")]
+        public string ChannelName { get; set; }
+
+        [JsonProperty("type")]
+        public string Type { get; set; }
+    }
+}
