+++ OutlookMatters/Mattermost/ChannelList.cs
@@ -7,16 +7,4 @@ namespace OutlookMatters.Mattermost
         [JsonProperty("channels")]
         public List<Channel> Channels { get; set; }
     }
-
-    public class Channel
-    {
-        [JsonProperty("id")]
-        public string ChannelId { get; set; }
-
-        [JsonProperty("display_name")]
-        public string ChannelName { get; set; }
-
-        [JsonProperty("type")]
-        public string Type { get; set; }
-    }
 }
